Added draft post cardview setting
