--- app/src/main/res/menu/draft_popup_setting.xml
@@ -1,13 +1,0 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:id="@+id/view"
-        android:title="View"
-        android:icon="@drawable/ic_share_wordpress"/>
-    <item android:id="@+id/publish"
-        android:title="Publish"/>
-    <item android:id="@+id/duplicate"
-        android:title="Duplicate"/>
-    <item android:id="@+id/share"
-        android:title="Share"/>
-    <item android:id="@+id/trash"
-        android:title="Trash"/>
-</menu>
